pyproject.toml
[project]
name = "langchain-tmp"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "gradio>=5.27.0",
    "langchain>=0.3.23",
    "langchain-cli>=0.0.36",
    "langchain-community>=0.3.21",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.31",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langserve>=0.3.1",
    "mcp[cli]>=1.8.0",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.0",
    "uvicorn>=0.34.1",
    "wikipedia>=1.4.0",
]
